requirements.txt
Flask==3.0.0
python-Levenshtein==0.25.0
Werkzeug==3.0.1
